requirements.txt
timm
lightning[pytorch-extra]
jsonargparse
einops
opencv-python
numpy
kornia
matplotlib==3.10.8
Pillow
mediapy
open3d
viser
ipykernel
jupyter
transformers>=4.37.0
accelerate
peft
bitsandbytes
sentencepiece
protobuf
scipy
scikit-image
mypy
ruff
black
gdown